Fixed compilation waring on win
git-svn-id: svn://falconpl.org/falcon/core/trunk@591 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/core_module/table.cpp
@@ -1303,7 +1303,7 @@ FALCON_FUNC  Table_getPage ( ::Falcon::VMachine *vm )
    }
 
    CoreTable *table = static_cast<CoreTable *>( vm->self().asObject()->getUserData() );
-   uint32 pos = i_pos == 0 ? table->currentPageId() : i_pos->forceInteger();
+   uint32 pos = i_pos == 0 ? table->currentPageId() : (uint32) i_pos->forceInteger();
    CoreArray* page = table->page( pos );
 
    if ( page == 0 )
